refactor: 가전제품 개별 컴포넌트 분리
