+++ src/app/_component/appliances/Appliance.module.css
@@ -1,0 +1,32 @@
+.CircleContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+
+  width: 69px;
+  height: 89px;
+}
+
+.Circle {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 6.9rem;
+  height: 6.9rem;
+  border-radius: 200%;
+
+  background-color: white;
+  border: 3px solid #83c63f;
+  box-sizing: border-box;
+}
+
+.CircleText {
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+  color: #000;
+  text-align: center;
+}
+
